requirements.txt
# ======================================
# Core libraries
# ======================================
PyYAML>=6.0

# ======================================
# Confluence API
# ======================================
atlassian-python-api>=3.41.0

# ======================================
# Rocket.Chat API
# ======================================
rocketchat-API>=1.30.0

# ======================================
# Data handling
# ======================================
pandas>=1.5.0

# ======================================
# Excel handling
# ======================================
openpyxl>=3.1.0
pywin32>=306 ; platform_system == "Windows"

# ======================================
# Web automation (Apps / Web勤怠)
# ======================================
selenium>=4.10.0
webdriver-manager>=4.0.0

# ======================================
# Others
# ======================================
requests>=2.31.0
urllib3>=2.0.0